Armstrong number completed..

+++ contracts/LabCycle/EvenNum.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.14;
 
 contract EvenNum{
 
-    function even(int n)public pure returns(int){
+    function even(int n)public pure returns(int8){
         if(n%2==0)
             return 1;
         else
